Move fixes code to a subdirectory.
