+++ src/Sarif.Viewer.VisualStudio/Fixes/FixSuggestedActionsSourceProvider.cs
@@ -7,7 +7,7 @@ using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.Utilities;
 
-namespace Microsoft.Sarif.Viewer
+namespace Microsoft.Sarif.Viewer.Fixes
 {
     /// <summary>
     /// Provides a <see cref="FixSuggestedActionsSource"/> for a specified <see cref="ITextView"/>
