+++ src/Sarif.Viewer.VisualStudio/Sarif.Viewer.VisualStudio.csproj
@@ -64,9 +64,9 @@
     <Compile Include="ErrorList\ErrorListCommand.cs" />
     <Compile Include="ErrorList\ISarifErrorListEventSelectionService.cs" />
     <Compile Include="ErrorList\SarifResultTableEntry.cs" />
-    <Compile Include="FixSuggestedAction.cs" />
-    <Compile Include="FixSuggestedActionsSource.cs" />
-    <Compile Include="FixSuggestedActionsSourceProvider.cs" />
+    <Compile Include="Fixes\FixSuggestedAction.cs" />
+    <Compile Include="Fixes\FixSuggestedActionsSource.cs" />
+    <Compile Include="Fixes\FixSuggestedActionsSourceProvider.cs" />
     <Compile Include="Services\ICloseSarifLogService.cs" />
     <Compile Include="Services\ILoadSarifLogService2.cs" />
     <Compile Include="Services\ILoadSarifLogService.cs" />
